App.cpp, GameMain.cpp のコンパイラ警告に対応。

+++ source/Main.cpp
@@ -34,7 +34,7 @@ int WINAPI WinMain( HINSTANCE hInst, HINSTANCE, LPSTR, int nCmdShow )
 		std::string message = std::string( "exception on " ) + e.file() + ":" + common::serialize( e.line() ) + "\n";
 
 		LPVOID output = 0;
-		FormatMessage( FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS | FORMAT_MESSAGE_ALLOCATE_BUFFER, 0, e.data(), MAKELANGID( LANG_NEUTRAL, SUBLANG_DEFAULT ), ( LPSTR )( & output ), 0, 0 );
+		FormatMessage( FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS | FORMAT_MESSAGE_ALLOCATE_BUFFER, 0, static_cast< DWORD >( e.data() ), MAKELANGID( LANG_NEUTRAL, SUBLANG_DEFAULT ), ( LPSTR )( & output ), 0, 0 );
 		message += static_cast< LPCTSTR >( output );
 
 		LocalFree( output );
